use simpleHandSignRecognizer
To remove dupcalited code
