+++ Assets/Scripts/Recognizer/SimpleHandSignRecognizer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 064eb97215b3d4891baeaff5db374c0b
+timeCreated: 1450243099
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
